Map receives player ships + changes to Wei's code
This is an incomplete commit of GameState and Map including some changes
to Map to allow the initialization of the ship arrays for both players.
Also changed the return types of 2 of Wei's methods (prepareMoveShip and
prepareTurnShip) and removed a whole bunch of fields. For reasons why
this was done, contact me.