GameState.toString() + GameStatePacket tested.
A test for the GameStatePacket has been written and the class has been
successfully tested in serializing and deserializing a whole GameState
into a byte[] array.

To test that the deserialized GameState object is identical to the old
one, GameState.toString() has been added to print the contents of the
game state to the console. This requires corresponding toString()
implementations in all classes contained within a GameState, which have
been implemented as well.