GUI create and join working.
Now players can vote on a coral reef and when both players accept the
game is started. Game start is not yet implemented, but //TODO tags are
placed on the lines where the game start code should be inserted.

Things that would be nice to be added soon:
No visual indication on the server side whether a player has connected
or not.
No visual indication on the other player's vote.